+++ src/Products/PloneMeeting/ToolPloneMeeting.py
@@ -448,6 +448,23 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
             user_groups = [plone_group for plone_group in user_groups if plone_group.startswith(org_uid)]
         return sorted(user_groups)
 
+    def group_is_not_empty_cachekey(method, self, org_uid, suffix, user_id=None):
+        '''cachekey method for self.group_is_not_empty.'''
+        date = get_cachekey_volatile('Products.PloneMeeting.ToolPloneMeeting.group_is_not_empty')
+        return (date,
+                self._users_groups_value(),
+                org_uid,
+                suffix,
+                user_id)
+
+    def group_is_not_empty(self, org_uid, suffix, user_id=None):
+        '''Is there any user in the group?'''
+        portal = api.portal.get()
+        plone_group_id = get_plone_group_id(org_uid, suffix)
+        # for performance reasons, check directly in source_groups stored data
+        group_users = portal.acl_users.source_groups._group_principal_map.get(plone_group_id, [])
+        return len(group_users) and not user_id or user_id in group_users
+
     def get_orgs_for_user_cachekey(method,
                                    self,
                                    user_id=None,
@@ -1105,7 +1122,21 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
                 if not destMeetingConfig.getKeepOriginalToPrintOfClonedItems():
                     newAnnex.to_print = \
                         get_category_object(newAnnex, newAnnex.content_category).to_print
-        # update annex index
+
+        # Change the proposing group if the item owner does not belong to
+        # the defined proposing group, except if p_keepProposingGroup is True
+        if not keepProposingGroup:
+            userGroupUids = self.get_orgs_for_user(
+                user_id=newOwnerId, suffixes=['creators', ], the_objects=False)
+            if userGroupUids and newItem.getProposingGroup(True) not in userGroupUids:
+                newItem.setProposingGroup(userGroupUids[0])
+
+        if newOwnerId != loggedUserId:
+            plone_utils = api.portal.get_tool('plone_utils')
+            plone_utils.changeOwnershipOf(newItem, newOwnerId)
+
+        # update annex index after every user/groups things are setup
+        # because annexes confidentiality relies on all this
         update_all_categorized_elements(newItem)
         # remove defered call to 'update_all_categorized_elements'
         self.REQUEST.set('defer_update_categorized_elements', False)
@@ -1124,18 +1155,6 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
         for ann in annotations:
             if ann.startswith(SENT_TO_OTHER_MC_ANNOTATION_BASE_KEY):
                 del annotations[ann]
-
-        # Change the proposing group if the item owner does not belong to
-        # the defined proposing group, except if p_keepProposingGroup is True
-        if not keepProposingGroup:
-            userGroupUids = self.get_orgs_for_user(
-                user_id=newOwnerId, suffixes=['creators', ], the_objects=False)
-            if userGroupUids and newItem.getProposingGroup(True) not in userGroupUids:
-                newItem.setProposingGroup(userGroupUids[0])
-
-        if newOwnerId != loggedUserId:
-            plone_utils = api.portal.get_tool('plone_utils')
-            plone_utils.changeOwnershipOf(newItem, newOwnerId)
 
         self.REQUEST.set('currentlyPastingItems', False)
         return newItem